requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
openai>=1.0.0
sentence-transformers
torch
openai-whisper
pydub
ffmpeg-python